requirements.txt
pymongo
kafka-python
Flask